hover grid imgs

+++ server/portfolio/models.py
@@ -3,17 +3,12 @@ import os
 
 class Work(models.Model):
     title = models.CharField(max_length=100, blank=True)
-    image = models.ImageField(upload_to='works/') 
-
+    image = models.ImageField(upload_to='works/')
+    
     def __str__(self):
         return self.title or "Untitled"
-    
 
-    #deleta as imagens do banco de dados e da pasta media tambem
     def delete(self, *args, **kwargs):
-        
-        if self.image:
-            if os.path.isfile(self.image.path):
-                os.remove(self.image.path)
-        
+        if self.image and os.path.isfile(self.image.path):
+            os.remove(self.image.path)
         super().delete(*args, **kwargs)
